弾の発射を実装 ・Spaceキーを押して発射 ・発射された弾は上に移動する オリジナルの機能 ・Vキーで射撃モードの切り替え(単発⇔3方向) ・射撃モードの表示 ・画面内に存在できる弾は3発まで(3方向モードを使っても4発以上の弾は存在できない)

+++ DirectX2DLibraryCpp/Src/Common/Enemy.h
@@ -11,8 +11,8 @@
 class Enemy {
 
 public:
-	double pos_x = WINDOW_WIDTH / 2 - 25; // 座標X
-	double pos_y = WINDOW_HEIGHT / 3; // 座標Y
+	float pos_x = WINDOW_WIDTH / 2 - 25; // 座標X
+	float pos_y = WINDOW_HEIGHT / 3; // 座標Y
 
 };
